chore: move handlers to lib_final.py
